app/requirements.txt
pillow
numpy
scipy
opencv-python
tensorflow
keras
keras_retinanet
